+++ apps/deploy-web/src/queries/useExactFeeAllowanceQuery.ts
@@ -1,13 +1,13 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { useAuthZService } from "@src/hooks/useAuthZService";
+import { useServices } from "@src/context/ServicesProvider";
 import { QueryKeys } from "@src/queries/queryKeys";
 
 export function useExactFeeAllowanceQuery(granter: string, grantee: string, { enabled = true } = {}) {
-  const allowanceHttpService = useAuthZService();
+  const { authzHttpService } = useServices();
   return useQuery({
     queryKey: QueryKeys.getFeeAllowancesKey(granter, grantee),
-    queryFn: () => allowanceHttpService.getFeeAllowanceForGranterAndGrantee(granter, grantee),
+    queryFn: () => authzHttpService.getFeeAllowanceForGranterAndGrantee(granter, grantee),
     enabled
   });
 }