+++ apps/deploy-web/src/queries/useGrantsQuery.ts
@@ -1,9 +1,9 @@
 import type { DepositDeploymentGrant } from "@akashnetwork/http-sdk";
 import type { UseQueryOptions } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
+import type { AxiosInstance } from "axios";
 
-import { useSettings } from "@src/context/SettingsProvider";
-import { useAuthZService } from "@src/hooks/useAuthZService";
+import { useServices } from "@src/context/ServicesProvider";
 import type { AllowanceType, PaginatedAllowanceType, PaginatedGrantType } from "@src/types/grant";
 import { ApiUrlService, loadWithPagination } from "@src/utils/apiUtils";
 import { QueryKeys } from "./queryKeys";
@@ -14,28 +14,26 @@ export function useGranterGrants(
   limit: number,
   options: Omit<UseQueryOptions<PaginatedGrantType>, "queryKey" | "queryFn"> = {}
 ) {
-  const { settings } = useSettings();
-  const allowanceHttpService = useAuthZService();
+  const { authzHttpService } = useServices();
   const offset = page * limit;
 
-  options.enabled = options.enabled !== false && !!address && !!settings.apiEndpoint;
+  options.enabled = options.enabled !== false && !!address && !!authzHttpService.defaults.baseURL;
 
   return useQuery({
     queryKey: QueryKeys.getGranterGrants(address, page, offset),
-    queryFn: () => allowanceHttpService.getPaginatedDepositDeploymentGrants({ granter: address, limit, offset }),
+    queryFn: () => authzHttpService.getPaginatedDepositDeploymentGrants({ granter: address, limit, offset }),
     ...options
   });
 }
 
 export function useGranteeGrants(address: string, options: Omit<UseQueryOptions<DepositDeploymentGrant[]>, "queryKey" | "queryFn"> = {}) {
-  const allowanceHttpService = useAuthZService();
-  const { settings } = useSettings();
+  const { authzHttpService } = useServices();
 
-  options.enabled = options.enabled !== false && !!address && !!settings.apiEndpoint;
+  options.enabled = options.enabled !== false && !!address && !!authzHttpService.defaults.baseURL;
 
   return useQuery({
     queryKey: QueryKeys.getGranteeGrants(address || "UNDEFINED"),
-    queryFn: () => allowanceHttpService.getAllDepositDeploymentGrants({ grantee: address, limit: 1000 }),
+    queryFn: () => authzHttpService.getAllDepositDeploymentGrants({ grantee: address, limit: 1000 }),
     ...options
   });
 }
@@ -46,31 +44,30 @@ export function useAllowancesIssued(
   limit: number,
   options: Omit<UseQueryOptions<PaginatedAllowanceType>, "queryKey" | "queryFn"> = {}
 ) {
-  const { settings } = useSettings();
-  const allowanceHttpService = useAuthZService();
+  const { authzHttpService } = useServices();
   const offset = page * limit;
 
-  options.enabled = options.enabled !== false && !!address && !!settings.apiEndpoint;
+  options.enabled = options.enabled !== false && !!address && !!authzHttpService.defaults.baseURL;
 
   return useQuery({
     queryKey: QueryKeys.getAllowancesIssued(address, page, offset),
-    queryFn: () => allowanceHttpService.getPaginatedFeeAllowancesForGranter(address, limit, offset),
+    queryFn: () => authzHttpService.getPaginatedFeeAllowancesForGranter(address, limit, offset),
     ...options
   });
 }
 
-async function getAllowancesGranted(apiEndpoint: string, address: string) {
-  return await loadWithPagination<AllowanceType[]>(ApiUrlService.allowancesGranted(apiEndpoint, address), "allowances", 1000);
+async function getAllowancesGranted(chainApiHttpClient: AxiosInstance, address: string) {
+  return await loadWithPagination<AllowanceType[]>(ApiUrlService.allowancesGranted("", address), "allowances", 1000, chainApiHttpClient);
 }
 
 export function useAllowancesGranted(address: string, options: Omit<UseQueryOptions<AllowanceType[]>, "queryKey" | "queryFn"> = {}) {
-  const { settings } = useSettings();
+  const { chainApiHttpClient } = useServices();
 
-  options.enabled = options.enabled !== false && !!address && !!settings.apiEndpoint;
+  options.enabled = options.enabled !== false && !!address && !!chainApiHttpClient.defaults.baseURL;
 
   return useQuery({
     queryKey: address ? QueryKeys.getAllowancesGranted(address) : [],
-    queryFn: () => getAllowancesGranted(settings.apiEndpoint, address),
+    queryFn: () => getAllowancesGranted(chainApiHttpClient, address),
     ...options
   });
 }